refactor PvpTournamentInstance
